+++ onsiteprint-plugin/acf-blocks/button-link.php
@@ -9,26 +9,30 @@
  *  @package WordPress
  *  @subpackage OnsitePrint Plugin
  *  @since OnsitePrint Plugin 1.0
- ?  Updated: 2022-10-28 - 14:01 (Y:m:d - H:i)
+ ?  Updated: 2025-10-19 - 02:18 (Y:m:d - H:i)
+ ?  Info: Fixed Variables
 
 ---------------------------------------------------------------------------
  #  The Block Data
 --------------------------------------------------------------------------- */
 
-$buttonLink = get_field('button_link');
-$buttonSize = get_field('button_size');
-$buttonStyle = get_field('button_style');
-$buttonColor = get_field('button_color');
-$buttonIcon = get_field('button_icon');
-$buttonIconPosition = get_field('button_icon_position');
-$buttonTitleVisibility = get_field('button_title_visibility');
-$buttonMaxWidthNumber = get_field('button_max_width_number');
-$buttonMaxWidthUnit = get_field('button_max_width_unit');
+$buttonLink = get_field('button_link') ?: false;
+$buttonSize = get_field('button_size') ?: false;
+$buttonStyle = get_field('button_style') ?: false;
+$buttonColor = get_field('button_color') ?: false;
+$buttonIcon = get_field('button_icon') ?: false;
+$buttonIconPosition = get_field('button_icon_position') ?: '';
+$buttonTitleVisibility = get_field('button_title_visibility') ?: 0;
+$buttonMaxWidthNumber = get_field('button_max_width_number') ?: false;
+$buttonMaxWidthUnit = get_field('button_max_width_unit') ?: 'px';
+$titleVisibility = '';
+$iconPosition = '';
+$icon = '';
 
 $id = 'op-' . $block['id'];
 
 if( ! empty( $block['anchor'] ) ) {
-	$id = $block['anchor'];
+    $id = $block['anchor'];
 }
 
 $className = 'op-button';
@@ -48,7 +52,7 @@ if ( $buttonColor ) {
 if( $buttonLink ) {
     //var_dump($buttonLink);
     $url = $buttonLink['url'];
-	$target = $buttonLink['target'];
+    $target = $buttonLink['target'];
     $title = $buttonLink['title'];
 
     if( $target ) {
@@ -56,16 +60,16 @@ if( $buttonLink ) {
     }
 
     if( ! $title || "" ) {
-        if ( substr($url, 0, 7) == 'http://' ) {
-            $title = substr($url, 7);
-        } else if ( substr($url, 0, 8) == 'https://' ) {
-            $title = substr($url, 8);
+        if ( substr( $url, 0, 7 ) == 'http://' ) {
+            $title = substr( $url, 7 );
+        } else if ( substr( $url, 0, 8 ) == 'https://' ) {
+            $title = substr( $url, 8 );
         } else {
             $title = $url;
         }
     }
 } else {
-    $url = 'https://onsiteprint.dk';
+    $url = site_url();
     $title = 'OnsitePrint';
 }
 
@@ -83,14 +87,6 @@ if( $buttonIconPosition ) {
 
 if ( $buttonTitleVisibility == 1 ) {
     $titleVisibility = ' data-title-visibility=' . $buttonTitleVisibility;
-    
-    if ( $buttonMaxWidthNumber ) {
-        $titleMaxWidth = ' style=max-width:' . $buttonMaxWidthNumber . $buttonMaxWidthUnit;
-    }
-} else {
-    if ( $buttonMaxWidthNumber ) {
-        $maxWidth = ' style=max-width:' . $buttonMaxWidthNumber . $buttonMaxWidthUnit;
-    }
 }
 
 
@@ -100,9 +96,9 @@ if ( $buttonTitleVisibility == 1 ) {
 --------------------------------------------------------------------------- */
 ?>
 
-<a id="<?= esc_attr($id) ?>" class="<?= esc_attr($className) ?>" href="<?= esc_attr($url) ?>" <?php echo esc_attr($target) . esc_attr($buttonColor) . esc_attr($icon) . esc_attr($iconPosition) . esc_attr($titleVisibility) . esc_attr($maxWidth) ?>>
-    <?php if( $buttonIcon ) { ?>
+<a id="<?= esc_attr( $id ) ?>" class="<?= esc_attr( $className ) ?>" href="<?= esc_attr( $url ) ?>" <?php echo esc_attr( $target ) . esc_attr( $buttonColor ) . esc_attr( $icon ) . esc_attr( $iconPosition ) . esc_attr( $titleVisibility )?>>
+    <?php if ( $buttonIcon ) { ?>
         <span class="op-icon" role="img" aria-label="<?= esc_attr($buttonIcon) ?> Icon"></span>
-    <? } ?>
-    <span class="op-button-title"<?= esc_attr($titleMaxWidth) ?>><?= esc_attr($title) ?></span>
-</a><!-- #<?= esc_attr($id) ?> -->
+    <?php } ?>
+    <span class="op-button-title"><?= esc_attr( $title ) ?></span>
+</a><!-- #<?= esc_attr( $id ) ?> -->